+++ src/ChatifyServiceProvider.php
@@ -90,6 +90,8 @@ class ChatifyServiceProvider extends ServiceProvider
             __DIR__ . '/assets/js' => public_path('js/chatify'),
             // Images
             __DIR__ . '/assets/imgs' => storage_path('app/public/' . $userAvatarFolder),
+             // CSS
+             __DIR__ . '/assets/sounds' => public_path('sounds/chatify'),
         ], 'chatify-assets');
     }